requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pandas
numpy
requests
faiss-cpu
sentence-transformers
torch
transformers
huggingface-hub